inference_server/requirements.txt
# Local dev inference server (FastAPI). Align ultralytics with aws_resources/sagemaker/requirements.txt where practical.
#
# To avoid very large CUDA wheels on Linux, install CPU PyTorch first, then the rest:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#   pip install -r requirements.txt
#
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
ultralytics>=8.3.0
Pillow>=10.1.0
numpy>=1.24.0
jinja2>=3.1.0
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
